organizing code with hooks
